Added error highlighting in edit.kid.
When there is an error in the markup text entered for an article, the location of the error is highlighted in the text box. Currently, this definitely won't work in IE, and hasn't been tested on anything except Firefox. Also, there is a separate bug in the markup parser leading to inaccurate line/column numbers.